Replace getOption("project_root") with here() call
I was setting the project root in the .Rprofile with here::here() but
it was pretty pointless since I was using getOption("project_root")
everywhere. This commit replaces all instances of
getOption("project_root") with here::here(). This also makes it play
more nicely with shiny server, since it starts from the directory
where the app.R file is located.